feat(web): mobile responsive
