+++ src/pages/Home/LoadingItems.tsx
@@ -3,7 +3,6 @@ import styled from 'styled-components'
 
 const Container = styled.div`
   display: flex;
-  height: 32px;
 `
 
 const LoadingImage = styled.img`
