+++ src/pages/Home/SubmitEntries/index.tsx
@@ -1,13 +1,15 @@
 import React from 'react'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
+import { landscapeStyle } from 'styles/landscapeStyle'
 import EarnRewards from './EarnRewards'
 import SuggestAndSubmit from './SuggestAndSubmit'
 
 const Container = styled.div`
   display: flex;
-  width: 80%;
+  flex-direction: column;
+  width: 84vw;
+  align-items: flex-start;
   margin-bottom: 24px;
-  align-items: center;
   border-radius: 12px;
   background: #380c65;
   padding: 16px 32px;
@@ -16,6 +18,16 @@ const Container = styled.div`
   justify-content: space-between;
   font-size: 22px;
   box-sizing: border-box;
+  flex-wrap: wrap;
+  gap: 16px;
+
+  ${landscapeStyle(
+    () => css`
+      width: 80%;
+      flex-direction: row;
+      align-items: center;
+    `
+  )}
 `
 
 interface ISubmitEntries {
@@ -30,12 +42,10 @@ const SubmitEntries: React.FC<ISubmitEntries> = ({
   curateContractAddress,
 }) => {
   return (
-    <>
-      <Container>
-        <EarnRewards />
-        <SuggestAndSubmit activeList={activeList} />
-      </Container>
-    </>
+    <Container>
+      <EarnRewards />
+      <SuggestAndSubmit activeList={activeList} />
+    </Container>
   )
 }
 export default SubmitEntries
